doc & refactor: added var doc and correct type.
To avoid messy int sizes, the size is now predetermined using stdint.
The doc for each variable is now provided (in french).

+++ src/Date.cpp
@@ -1,10 +1,11 @@
 #include "Date.h"
-#include "Arduino.h"
 
-char* substring(const char* _buffer_value, const int start){
+
+char* substring(const char* _buffer_value, const uint8_t start){
     char str[2];
-    int i_buffer = start;
-    for(int i_str = 0; i_str < 2; i_str++){
+    uint8_t i_buffer = start;
+    
+    for(uint8_t i_str = 0; i_str < 2; i_str++){
         str[i_str] = _buffer_value[i_buffer + i_str];
         i_str++;
     }
@@ -17,6 +18,7 @@ void Date::parse(const char* horodate){
     annee = atoi(substring(horodate, 1));
     mois = atoi(substring(horodate, 3));
     jours = atoi(substring(horodate, 5));
+
     heures = atoi(substring(horodate, 7));
     minutes = atoi(substring(horodate, 9));
     secondes = atoi(substring(horodate, 11));
